Move the map to its own file and discourage use
